requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.0.11
defusedxml==0.6.0
entrypoints==0.3
filetype==1.0.5
Flask==1.1.1
Flask-BasicAuth==0.2.0
idna==2.8
igramscraper==0.3.2
ImageHash==4.0
imageio==2.1.2
imbalanced-learn==0.6.1
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
moviepy==0.2.3.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.18.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
peewee==3.13.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.1
python-slugify==3.0.2
pytz==2019.3
PyWavelets==1.1.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.21.0
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.13.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tornado==6.0.3
tqdm==4.11.2
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
